LayoutTests/fast/repaint/hidpi-content-inside-iframe-leaves-trails-expected.txt

PASS repaintRects.indexOf('rect 0 0 101.50 100') is not -1
PASS successfullyParsed is true

TEST COMPLETE
